crawler/requirements.txt
requests==2.30.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.6
schedule==1.2.0
selenium==4.10.0